deleted NfcActivity and moved all the methods back to the MainActivity, no problem with the server and no need to the timer now.
